pyproject.toml
[project]
name = "autoresearch"
version = "0.1.0"
description = "Autonomous pretraining research swarm"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kernels>=0.11.7; sys_platform != 'darwin'",
    "matplotlib>=3.10.8",
    "mlx>=0.30.0; sys_platform == 'darwin'",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
    "requests>=2.32.0",
    "rustbpe>=0.1.0",
    "tiktoken>=0.11.0",
    "python-docx>=1.1.0",
    "torch==2.9.1; sys_platform != 'darwin'",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform != 'darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
